// backend/src/modules/pickupSlot/models/pickupSlot.model.d.ts
import type { IPickupSlot } from "../interfaces/IPickupSlot.js";
export declare const PickupSlot: import("mongoose").Model<IPickupSlot, {}, {}, {}, import("mongoose").Document<unknown, {}, IPickupSlot, {}, import("mongoose").DefaultSchemaOptions> & IPickupSlot & Required<{
    _id: import("mongoose").Types.ObjectId;
}> & {
    __v: number;
} & {
    id: string;
}, any, IPickupSlot>;
//# sourceMappingURL=pickupSlot.model.d.ts.map